requirements.txt
Django==5.0.3
psycopg==3.1.18
djangorestframework==3.15.0
pytest-django==4.8.0
model-bakery==1.17.0
